dev 2020.07.01.1

+++ LSLRecorder/src/DataStream/OutputDataFile/DataBlock/ByteBlock.java
@@ -23,7 +23,7 @@ import edu.ucsd.sccn.LSL;
 
 public class ByteBlock extends DataBlock<Byte>
 {
-	public ByteBlock(int seqNum, String name, int nCols, Byte[] data) 
+	public ByteBlock(int seqNum, String name, long nCols, Byte[] data) 
 	{
 		super( seqNum, name, LSL.ChannelFormat.int8, nCols, data);
 	}
